Performs project templating
