+++ templates/block_link.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * View Link post
+ * @var string Url site
+ */
+?>
+<div class="post-link__wrapper">
+    <a class="post-link__external" href="http://" title="Перейти по ссылке">
+        <div class="post-link__info-wrapper">
+            <div class="post-link__icon-wrapper">
+                <img src="https://www.google.com/s2/favicons?domain=vitadental.ru" alt="Иконка">
+            </div>
+            <div class="post-link__info">
+                <h3><?= htmlspecialchars($content); ?></h3>
+            </div>
+        </div>
+        <span><?= htmlspecialchars($content); ?></span>
+    </a>
+</div>
