+++ templates/block_photo.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * View Photo post
+ * @var string Url photo
+ */
+?>
+<div class="post-photo__image-wrapper">
+    <img src="img/<?= htmlspecialchars($content); ?>" alt="Фото от пользователя" width="360" height="240">
+</div>
