+++ templates/block_quote.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * View Quote post
+ * @var string Text content
+ */
+?>
+<blockquote>
+    <p>
+        <?= htmlspecialchars($content); ?>
+    </p>
+    <cite>Неизвестный Автор</cite>
+</blockquote>
