+++ templates/layout.php
@@ -1,10 +1,15 @@
+<?php
+/**
+ * Main layout
+ */
+?>
 <!DOCTYPE html>
 <html lang="ru">
 <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>readme: популярное</title>
+    <title><?= $title_page; ?></title>
     <link rel="stylesheet" href="css/main.css">
 </head>
 <body class="page">
@@ -110,7 +115,7 @@
     </div>
 </header>
 
-<!-- content -->
+<?= $content; ?>
 
 <footer class="footer">
     <div class="footer__wrapper">
