+++ templates/main.php
@@ -1,3 +1,10 @@
+<?php
+/**
+ * View page popular posts
+ */
+require_once('helpers.php');
+
+?>
 <section class="page__main page__main--popular">
     <div class="container">
         <h1 class="page__title page__title--popular">Популярное</h1>
@@ -91,56 +98,9 @@
                         <h2><?=$post['title']; ?></h2>
                     </header>
                     <div class="post__main">
-                        <!--здесь содержимое карточки-->
-                        <?php if($post['type'] === 'quote'): ?>
-                            <!--содержимое для поста-цитаты-->
-                            <blockquote>
-                                <p>
-                                    <?=$post['content']; ?>
-                                </p>
-                                <cite>Неизвестный Автор</cite>
-                            </blockquote>
-                        <?php elseif($post['type'] === 'link'): ?>
-                            <!--содержимое для поста-ссылки-->
-                            <div class="post-link__wrapper">
-                                <a class="post-link__external" href="http://" title="Перейти по ссылке">
-                                    <div class="post-link__info-wrapper">
-                                        <div class="post-link__icon-wrapper">
-                                            <img src="https://www.google.com/s2/favicons?domain=vitadental.ru" alt="Иконка">
-                                        </div>
-                                        <div class="post-link__info">
-                                            <h3><?=$post['content']; ?></h3>
-                                        </div>
-                                    </div>
-                                    <span><?=$post['content']; ?></span>
-                                </a>
-                            </div>
-                        <?php elseif($post['type'] === 'photo'): ?>
-                            <!--содержимое для поста-фото-->
-                            <div class="post-photo__image-wrapper">
-                                <img src="img/<?=$post['content']; ?>" alt="Фото от пользователя" width="360" height="240">
-                            </div>
-                        <?php elseif($post['type'] === 'video'): ?>
-                            <!--содержимое для поста-видео-->
-                            <div class="post-video__block">
-                                <div class="post-video__preview">
-                                    <?=embed_youtube_cover($post['content']); ?>
-                                    <img src="img/coast-medium.jpg" alt="Превью к видео" width="360" height="188">
-                                </div>
-                                <a href="post-details.html" class="post-video__play-big button">
-                                    <svg class="post-video__play-big-icon" width="14" height="14">
-                                        <use xlink:href="#icon-video-play-big"></use>
-                                    </svg>
-                                    <span class="visually-hidden">Запустить проигрыватель</span>
-                                </a>
-                            </div>
-                        <?php elseif($post['type'] === 'text'): ?>
-                            <!--содержимое для поста-текста-->
-                            <p><?=cut_text($post['content']); ?></p>
-                            <?php if (strlen(cut_text($post['content'])) !== strlen($post['content'])): ?>
-                                <a class="post-text__more-link" href="#">Читать далее</a>
-                            <?php endif; ?>
-                        <?php endif; ?>
+                        <?= include_template('block_' . $post['type'] . '.php', [
+                            'content' => $post['content'],
+                        ]);?>
                     </div>
                     <footer class="post__footer">
                         <div class="post__author">
